--- src/main/java/com/haru/api/domain/term/service/TermService.java
@@ -1,8 +1,0 @@
-package com.haru.api.domain.term.service;
-
-import com.haru.api.domain.term.dto.TermResponseDTO;
-import com.haru.api.domain.term.entity.enums.TermType;
-
-public interface TermService {
-    TermResponseDTO.TermDetail getTermByType(TermType type);
-}
