+++ src/main/java/com/haru/api/term/application/service/TermUseCaseImpl.java
@@ -1,10 +1,11 @@
-package com.haru.api.domain.term.service;
+package com.haru.api.term.application.service;
 
-import com.haru.api.domain.term.converter.TermConverter;
-import com.haru.api.domain.term.dto.TermResponseDTO;
-import com.haru.api.domain.term.entity.Term;
-import com.haru.api.domain.term.entity.enums.TermType;
-import com.haru.api.domain.term.repository.TermRepository;
+import com.haru.api.term.application.converter.TermConverter;
+import com.haru.api.term.application.port.in.TermUseCase;
+import com.haru.api.term.presentation.dto.TermResponseDTO;
+import com.haru.api.term.domain.Term;
+import com.haru.api.term.domain.enums.TermType;
+import com.haru.api.term.infrastructure.TermRepository;
 import com.haru.api.global.apiPayload.code.status.ErrorStatus;
 import com.haru.api.global.apiPayload.exception.GeneralException;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class TermServiceImpl implements TermService {
+public class TermUseCaseImpl implements TermUseCase {
 
     private final TermRepository termsRepository;
